+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/auth/AuthInputConsumerIT.kt
@@ -68,7 +68,6 @@ class AuthInputConsumerIT @Autowired constructor(
         ),
         relation = RelationAuthExpectations(
             get = AuthExpectationType.Authorized,
-            post = AuthExpectationType.Forbidden,
             put = AuthExpectationType.Forbidden
         )
     )